move utility method to Calamari Common
